requirements.txt
torch>=2.0.0
transformers>=4.35.0
numpy>=1.21.0
datasets>=2.10.0
accelerate>=0.24.0
tqdm>=4.64.0
pyyaml>=6.0
sentence-transformers>=2.2.0

